Refactor simulation package
