+++ simulations/deck_calculator/starter.py
@@ -1,9 +1,4 @@
-import os
-import sys
-
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..")))
-
-from simulations.deck_calculator.adapter import SphereDeckCalculator
+from simulations.sphere_space_station_simulations import SphereDeckCalculator
 
 
 def main() -> None:
