+++ simulations/sphere_space_station_simulations/README.md
@@ -1,3 +1,11 @@
 # Sphere Space Station Simulations Library
 
-This package bundles the core calculations, animation helpers and data export functionality used by the example simulations.  The modules here can be reused by the thin adapters located in the individual simulation folders.
+This package bundles the core calculations, animation helpers and data export functionality used by the example simulations.  It is organised into subpackages:
+
+```
+geometry/   - core classes such as ``SphereDeckCalculator``
+reporting/  - CSV/HTML export utilities
+animation/  - functions producing Matplotlib animations
+```
+
+The modules here can be reused by the thin adapters located in the individual simulation folders.
