+++ simulations/sphere_space_station_simulations/__init__.py
@@ -1,1 +1,5 @@
-from .deck_calculator import SphereDeckCalculator
+"""Core simulation package for the Sphere Station."""
+
+from .geometry.deck import SphereDeckCalculator
+
+__all__ = ["SphereDeckCalculator"]
