+++ simulations/sphere_space_station_simulations/animation/README.md
@@ -1,0 +1,3 @@
+# Animation Package
+
+Functions to generate 3D Matplotlib animations from deck data.
