+++ simulations/sphere_space_station_simulations/animation/__init__.py
@@ -1,0 +1,5 @@
+"""Animation helper functions for deck calculations."""
+
+from . import deck
+
+__all__ = ["deck"]
