+++ simulations/sphere_space_station_simulations/geometry/README.md
@@ -1,0 +1,3 @@
+# Geometry Package
+
+Contains core geometry classes and calculations used by the simulation modules.
