+++ simulations/sphere_space_station_simulations/reporting/__init__.py
@@ -1,0 +1,5 @@
+"""Reporting helper functions for deck calculations."""
+
+from . import deck
+
+__all__ = ["deck"]
